Support for nightly builds

+++ jenkins-scripts/lib/debbuild-base.bash
@@ -13,6 +13,11 @@ if [ -z $RELEASE_REPO_DIRECTORY ]; then
     RELEASE_REPO_DIRECTORY=ubuntu
 fi;
 
+NIGHTLY_MODE=false
+if [ "${VERSION}" = "nightly" ]; then
+    NIGHTLY_MODE=true
+fi
+
 ###################################################
 # Boilerplate.
 # DO NOT MODIFY
@@ -111,31 +116,56 @@ pbuilder-dist $DISTRO $ARCH create --othermirror "deb http://packages.ros.org/ro
 # Step 0: Clean up
 rm -rf $WORKSPACE/build
 mkdir -p $WORKSPACE/build
-
-# Step 1: Get the source tarball
 cd $WORKSPACE/build
-wget --quiet -O ${PACKAGE_ALIAS}_$VERSION.orig.tar.bz2 $SOURCE_TARBALL_URI
 
-# Step 3: unpack tarball
-rm -rf $PACKAGE-$VERSION
-tar xf ${PACKAGE_ALIAS}_$VERSION.orig.tar.bz2
-cd $PACKAGE-$VERSION
+# Step 1: Get the source (nightly builds or tarball)
+if ${NIGHTLY_MODE}; then
+  apt-get install -y mercurial
+  hg clone https://bitbucket.org/osrf/$PACKAGE
+  PACKAGE_SRC_BUILD_DIR=$PACKAGE
+else
+  wget --quiet -O ${PACKAGE_ALIAS}_$VERSION.orig.tar.bz2 $SOURCE_TARBALL_URI
+  rm -rf $PACKAGE-$VERSION
+  tar xf ${PACKAGE_ALIAS}_$VERSION.orig.tar.bz2
+  PACKAGE_SRC_BUILD_DIR=$PACKAGE-$VERSION
+fi
 
 # Step 4: add debian/ subdirectory with necessary metadata files to unpacked source tarball
 rm -rf /tmp/$PACKAGE-release
 hg clone https://bitbucket.org/osrf/$PACKAGE-release /tmp/$PACKAGE-release
 cd /tmp/$PACKAGE-release
+# In nightly get the default latest version from default changelog
+if $NIGHTLY_MODE; then
+    # TODO: remove check when multidistribution reach default branch
+    if [ -f "${RELEASE_REPO_DIRECTORY}/debian/changelog" ]; then
+      UPSTREAM_VERSION=\$( sed -n '/(/,/)/ s/.*(\([^)]*\)).*/\1 /p' ${RELEASE_REPO_DIRECTORY}/debian/changelog | head -n 1 | tr -d ' ')
+    else
+      UPSTREAM_VERSION=\$( sed -n '/(/,/)/ s/.*(\([^)]*\)).*/\1 /p' ubuntu/debian/changelog | head -n 1 | tr -d ' ')
+    fi
+fi
 hg up $RELEASE_REPO_BRANCH
 
 # Adding extra directories to code. debian has no problem but some extra directories 
 # handled by symlinks (like cmake) in the repository can not be copied directly. 
 # Need special care to copy, using first a --dereference
-cd $WORKSPACE/build/$PACKAGE-$VERSION
+cd $WORKSPACE/build/\$PACKAGE_SRC_BUILD_DIR
 cp -a --dereference /tmp/$PACKAGE-release/${RELEASE_REPO_DIRECTORY}/* .
+
+# [nightly] Adjust version in nightly mode
+if $NIGHTLY_MODE; then
+  TIMESTAMP=\$(date '+%Y%m%d')
+  RELEASE_DATE=\$(date '+%a, %d %B %Y %T -0700')
+  REV=\$(hg parents --template="{node|short}\n")
+  NIGHTLY_VERSION_SUFFIX=\${UPSTREAM_VERSION}~hg\${TIMESTAMP}r\${REV}-${RELEASE_VERSION}
+  # Fix the changelog
+  sed -i -e "s/xxxxx/\${NIGHTLY_VERSION_SUFFIX}/g" debian/changelog
+  sed -i -e "s/ddddd/\${RELEASE_DATE}/g" debian/changelog
+  # TODO: Fix CMakeLists.txt ?
+fi
 
 # Step 5: use debuild to create source package
 #TODO: create non-passphrase-protected keys and remove the -uc and -us args to debuild
-debuild -S -uc -us --source-option=--include-binaries
+debuild --no-tgz-check -S -uc -us --source-option=--include-binaries
 
 if [ $PACKAGE != gazebo ]; then
   if [ $DISTRO = quantal ]; then
@@ -161,8 +191,14 @@ pbuilder-dist $DISTRO $ARCH build ../*.dsc
 sudo apt-get install -y openssh-client
 cd /var/packages/gazebo/ubuntu
 
-PKG_NAME=${PACKAGE_ALIAS}_${VERSION}-${RELEASE_VERSION}~${DISTRO}_${ARCH}.deb
-DBG_PKG_NAME=${PACKAGE_ALIAS}-dbg_${VERSION}-${RELEASE_VERSION}~${DISTRO}_${ARCH}.deb
+# Set proper package names
+if $NIGHTLY_MODE; then
+  PKG_NAME=${PACKAGE_ALIAS}_\${NIGHTLY_VERSION_SUFFIX}_${ARCH}.deb
+  DBG_PKG_NAME=${PACKAGE_ALIAS}-dbg_\${NIGHTLY_VERSION_SUFFIX}_${ARCH}.deb
+else
+  PKG_NAME=${PACKAGE_ALIAS}_${VERSION}-${RELEASE_VERSION}~${DISTRO}_${ARCH}.deb
+  DBG_PKG_NAME=${PACKAGE_ALIAS}-dbg_${VERSION}-${RELEASE_VERSION}~${DISTRO}_${ARCH}.deb
+fi
 
 MAIN_PKGS="/var/lib/jenkins/pbuilder/${DISTRO}-${ARCH}_result/\${PKG_NAME} /var/lib/jenkins/pbuilder/${DISTRO}_result/\${PKG_NAME}"
 DEBUG_PKGS="/var/lib/jenkins/pbuilder/${DISTRO}-${ARCH}_result/\${DBG_PKG_NAME} /var/lib/jenkins/pbuilder/${DISTRO}_result/\${DBG_PKG_NAME}"
@@ -172,7 +208,7 @@ for pkg in \${MAIN_PKGS}; do
     echo "looking for \$pkg"
     if [ -f \${pkg} ]; then
         echo "found \$pkg"
-        # Check for correctly generated packages size > 3Kb
+	# Check for correctly generated packages size > 3Kb
         test -z \$(find \$pkg -size +3k) && exit 1
         GNUPGHOME=$WORKSPACE/gnupg reprepro includedeb $DISTRO \${pkg}
         scp -o StrictHostKeyChecking=no -i $WORKSPACE/id_rsa \${pkg} ubuntu@gazebosim.org:/var/www/assets/distributions
